fix: client 환경에 세팅해야하는 옵션들 정리
